requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
browser-use==0.1.40
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
Flask==3.1.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-anthropic==0.3.3
langchain-core==0.3.51
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.30
markdownify==0.14.1
MarkupSafe==3.0.2
monotonic==1.6
narwhals==1.34.1
nest-asyncio==1.6.0
numpy==2.2.4
ollama==0.4.7
openai==1.73.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.1
playwright==1.51.0
posthog==3.24.1
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pyee==12.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.44.1
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
zstandard==0.23.0
aiofiles==24.1.0

PyMuPDF
ace_tools==0.0
ace_tools_open==0.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
executing==2.2.0
faiss-cpu==1.10.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
itables==2.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.55
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.31
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mpmath==1.3.0
msedge-selenium-tools==3.141.4
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
openai==1.76.0
orjson==3.10.16
ormsgpack==1.9.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pi==0.1.2
pillow==11.2.1
platformdirs==4.3.7
playwright==1.51.0
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydot==3.0.4
pyee==11.1.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
pyperclip==1.9.0
pyppeteer==2.0.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==3.141.0
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.14.0
tavily-python==0.7.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
websockets==10.4
widgetsnbextension==4.0.14
wsproto==1.2.0
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
